requirements.txt
pick==2.2.0
pyserial==3.5
